+++ opentech/apply/dashboard/tests/test_views.py
@@ -1,6 +1,6 @@
 from opentech.apply.funds.tests.factories import ApplicationSubmissionFactory, ApplicationRevisionFactory
 from opentech.apply.users.tests.factories import UserFactory, StaffFactory
-from opentech.apply.utils.tests import BaseViewTestCase
+from opentech.apply.utils.tests.tests import BaseViewTestCase
 
 
 class TestApplicantDashboard(BaseViewTestCase):
